Updates to the Hackceler8 2025 game code.
The game will be running on the console inside a custom cartridge which provides
* Extra controller slots for 4 player support
* A network port for comms with a server to send scoreboard info and persist game state (such as flags received) between console resets
* RNG support
* A team ID which is used to load team-specific graphics

The cartridge features are made available in-game by reading/writing special memory mapped IO addresses. When ran on emulators, the game falls back to defaults (e.g. using a PRNG, using the EA multitap for 4 player mode).

Apart from the cartridge features, this commit also adds,
* More flashy boss defeat sequence
* Team specific player graphics (with placeholder graphics for now)
* Misc code/level refactors